refactor(chat): extract route handler into focused modules
break the 1352-line monolithic POST handler into four modules:
- normalize-scope.ts: resolves request params into unified scope
- rag-context.ts: retrieval metadata, source dedup, fallback replies
- build-stream.ts: tool definitions, parent hints, LLM call assembly
- stream-events.ts: typed SSE event emission helpers

the route file is now a thin orchestrator (~336 lines) that calls
these extracted helpers. all behavior and error paths preserved.
